+++ Hearthstone Deck Tracker/HsReplay/Utility/HsReplayInfo.cs
@@ -1,8 +1,13 @@
-using System.Windows.Forms;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
+using Hearthstone_Deck_Tracker.Annotations;
+
 namespace Hearthstone_Deck_Tracker.HsReplay.Utility
 {
-	public class HsReplayInfo
+	public class HsReplayInfo : INotifyPropertyChanged
 	{
+		private string _uploadId;
+
 		public HsReplayInfo()
 		{
 			
@@ -13,7 +18,16 @@ namespace Hearthstone_Deck_Tracker.HsReplay.Utility
 			UploadId = uploadId;
 		}
 
-		public string UploadId { get; set; }
+		public string UploadId
+		{
+			get => _uploadId;
+			set
+			{
+				_uploadId = value; 
+				OnPropertyChanged();
+				OnPropertyChanged(nameof(Uploaded));
+			}
+		}
 
 		public int UploadTries { get; set; }
 
@@ -34,5 +48,13 @@ namespace Hearthstone_Deck_Tracker.HsReplay.Utility
 		public string ReplayUrl { get; set; }
 
 		public void UploadTry() => UploadTries++;
+
+		public event PropertyChangedEventHandler PropertyChanged;
+
+		[NotifyPropertyChangedInvocator]
+		protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
+		{
+			PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+		}
 	}
 }