+++ Hearthstone Deck Tracker/Windows/MainWindowControls/DeckChartsView.xaml.cs
@@ -1,0 +1,14 @@
+﻿using Hearthstone_Deck_Tracker.Hearthstone;
+
+namespace Hearthstone_Deck_Tracker.Windows.MainWindowControls
+{
+	public partial class DeckChartsView
+	{
+		public DeckChartsView()
+		{
+			InitializeComponent();
+		}
+
+		public void SetDeck(Deck deck) => ((DeckChartsViewModel)DataContext).Deck = deck;
+	}
+}